test(server): cover admin SCIM token UI and split audit API tests out
The three browser handlers for /admin/scim-tokens had no tests — all of
the old mixed file's tests exercised the JSON API. Add ten: the
AdminPage redirect ladder (no session, non-admin, unverified admin),
the rendered token list, the create form's one-time token display and
its flash redirects for invalid input, and revocation including the
cross-org case, where a foreign token behaves like a missing one and
survives.

Move handlers/api/org/audit.rs's 660-line inline test module to a
sibling audit/tests.rs per the test placement convention (~500-line
cutoff); the 25 tests are unchanged.
